requirements.txt
# Dépendances principales
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# NLP et IA
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# Traitement audio
SpeechRecognition==3.10.0
gTTS==2.3.2
pydub==0.25.1

# Tests
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Sécurité
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==41.0.5
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
secure==0.3.0
bleach==6.1.0
django-cors-headers==4.3.1

# Monitoring et logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilitaires
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6 